+++ HISTORY.md
@@ -34,6 +34,7 @@
 * objects: enabled custom categories for the stage
 * store: load user defined palettes
 * gui: restore custom palettes when activating a scene
+* added new category feature to palette context menu
 
 ### 2021-07-21
 * user defined custom block palettes, under construction